Refactor: 카카오 api 결과 dto로 받음
