+++ src/main/java/gift/Application.java
@@ -1,6 +1,6 @@
 package gift;
 
-import gift.dto.KakaoProperties;
+import gift.dto.kakao.KakaoProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
