+++ src/main/java/gift/dto/kakao/KakaoAccessTokenResponse.java
@@ -1,0 +1,12 @@
+package gift.dto.kakao;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record KakaoAccessTokenResponse(
+        String tokenType,
+        String accessToken,
+        String scope
+) {
+}
