+++ src/main/java/gift/dto/kakao/KakaoMessageRequest.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
