+++ src/main/java/gift/dto/kakao/KakaoUserInfoResponse.java
@@ -1,0 +1,10 @@
+package gift.dto.kakao;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record KakaoUserInfoResponse(
+        Long id
+) {
+}
